requirements.txt
#Flask==1.1.4
Flask
beautifulsoup4==4.10.0
requests==2.26.0
bleach==4.1.0
python-dateutil==2.8.2
lxml
